data-provider.js fix
As it is clearly obvious in the source code of data-provider.js (lines 26-49), only the observed property names from the istsos demo are taken into consideration. If the actual uri definitions are different than those in demo data, the code leaves the "urn" variable empty.
As the code at this point does not have any convenient way to convert the Observed property name into the full uri used I propose to modify the code so the whole uri is needed in the configuration.